update() : check session user id when call API

+++ app/api/uAPI/route.js
@@ -1,7 +1,7 @@
 import { getServerSession } from "next-auth/next";
 import { authOptions } from "@/app/api/auth/[...nextauth]/route";
 import { connectMongoDB } from "@/lib/mongodb";
-import API from "@/models/API"
+import API from "@/models/API";
 import mongoose from "mongoose";
 
 export async function GET(req) {
@@ -32,6 +32,20 @@ export async function GET(req) {
           status: 404,
         });
       }
+
+      const existingChannel = await Channel.findById(api.channel_id.toString());
+      if (!existingChannel) {
+        return new Response(JSON.stringify({ message: "Channel not found." }), {
+          status: 404,
+        });
+      }
+      if (session.user._id && session.user._id !== existingChannel.user_id) {
+        return new Response(
+          JSON.stringify({ message: "No access this Channel" }),
+          { status: 400 }
+        );
+      }
+
       return new Response(
         JSON.stringify({
           status: {
@@ -45,6 +59,19 @@ export async function GET(req) {
         }
       );
     } else {
+      const existingChannel = await Channel.findById(channel_id);
+      if (!existingChannel) {
+        return new Response(JSON.stringify({ message: "Channel not found." }), {
+          status: 404,
+        });
+      }
+      if (session.user._id && session.user._id !== existingChannel.user_id) {
+        return new Response(
+          JSON.stringify({ message: "No access this Channel" }),
+          { status: 400 }
+        );
+      }
+
       const filter = {
         ...(channel_id && {
           channel_id: new mongoose.Types.ObjectId(channel_id),
@@ -108,7 +135,8 @@ export async function POST(req) {
       api_auth,
       keywords,
     } = await req.json();
-    console.log(name,
+    console.log(
+      name,
       method_type,
       description,
       api_endpoint,
@@ -117,12 +145,26 @@ export async function POST(req) {
       api_headers,
       api_body,
       api_auth,
-      keywords);
+      keywords
+    );
 
     if (!name || !method_type || !api_endpoint || !channel_id || !keywords) {
       return new Response(JSON.stringify({ message: "API not found." }), {
         status: 404,
       });
+    }
+
+    const existingChannel = await Channel.findById(channel_id);
+    if (!existingChannel) {
+      return new Response(JSON.stringify({ message: "Channel not found." }), {
+        status: 404,
+      });
+    }
+    if (session.user._id && session.user._id !== existingChannel.user_id) {
+      return new Response(
+        JSON.stringify({ message: "No access this Channel" }),
+        { status: 400 }
+      );
     }
 
     const newAPI = new API({
@@ -140,15 +182,15 @@ export async function POST(req) {
     console.log("newAPI", newAPI);
     const savedAPI = await newAPI.save();
     return new Response(
-        JSON.stringify({
-          status: {
-            code: 200,
-            description: "Success create channel!!",
-          },
-          API: savedAPI,
-        }),
-        { status: 201 }
-      );
+      JSON.stringify({
+        status: {
+          code: 200,
+          description: "Success create channel!!",
+        },
+        API: savedAPI,
+      }),
+      { status: 201 }
+    );
   } catch (error) {
     return new Response(
       JSON.stringify({
@@ -172,50 +214,65 @@ export async function PUT(req) {
 
   try {
     const body = await req.json();
-    const { id, ...updateData } = body;
+    const {
+      id,
+      name,
+      method_type,
+      description,
+      api_endpoint,
+      channel_id,
+      api_params,
+      api_headers,
+      api_body,
+      api_auth,
+      keywords,
+    } = body;
+
     if (!id) {
       return new Response(JSON.stringify({ message: "API not found." }), {
         status: 404,
       });
     }
 
-    // const {
-    //   name,
-    //   method_type,
-    //   description,
-    //   api_endpoint,
-    //   channel_id,
-    //   api_params,
-    //   api_headers,
-    //   api_body,
-    //   api_auth,
-    //   keywords,
-    // } = await req.json();
+    const existingAPI = await API.findById(id);
+    if (!existingAPI) {
+      return new Response(JSON.stringify({ message: "API not found." }), {
+        status: 404,
+      });
+    }
 
-    // if (!name || !method_type || !api_endpoint || !channel_id || !keywords) {
-    //   return new Response(JSON.stringify({ message: "API not found." }), {
-    //     status: 404,
-    //   });
-    // }
+    const existingChannel = await Channel.findById(
+      existingAPI.channel_id.toString()
+    );
 
-    // const updatedAPI = await API.findByIdAndUpdate(id, {
-    //   name,
-    //   method_type,
-    //   description,
-    //   api_endpoint,
-    //   channel_id: new mongoose.Types.ObjectId(channel_id),
-    //   api_params,
-    //   api_headers,
-    //   api_body,
-    //   api_auth,
-    //   keywords,
-    // }, {
-    //   new: true,
-    // });
-
-    const updatedAPI = await API.findByIdAndUpdate(id, updateData, {
-      new: true,
-    });
+    if (!existingChannel) {
+      return new Response(JSON.stringify({ message: "Channel not found." }), {
+        status: 404,
+      });
+    }
+    if (session.user._id && session.user._id !== existingChannel.user_id) {
+      return new Response(
+        JSON.stringify({ message: "No access this Channel" }),
+        { status: 400 }
+      );
+    }
+    const updatedAPI = await API.findByIdAndUpdate(
+      id,
+      {
+        name,
+        method_type,
+        description,
+        api_endpoint,
+        api_params,
+        api_headers,
+        api_body,
+        api_auth,
+        keywords,
+      },
+      {
+        new: true,
+      }
+    );
 
     if (!updatedAPI) {
       return new Response(JSON.stringify({ message: "API not found." }), {
@@ -233,7 +290,6 @@ export async function PUT(req) {
       }),
       { status: 200 }
     );
-
   } catch (error) {
     return new Response(
       JSON.stringify({
@@ -264,13 +320,26 @@ export async function DELETE(req) {
       });
     }
 
-    const deletedAPI = await API.findByIdAndDelete(id);
-
-    if (!deletedAPI) {
+    const existingAPI = await API.findById(id);
+    if (!existingAPI) {
       return new Response(JSON.stringify({ message: "API not found." }), {
         status: 404,
       });
     }
+    const existingChannel = await Channel.findById(existingAPI.channel_id.toString());
+    if (!existingChannel) {
+      return new Response(JSON.stringify({ message: "Channel not found." }), {
+        status: 404,
+      });
+    }
+    if (session.user._id && session.user._id !== existingChannel.user_id) {
+      return new Response(
+        JSON.stringify({ message: "No access this Channel" }),
+        { status: 400 }
+      );
+    }
+
+    await API.findByIdAndDelete(id);
 
     return new Response(
       JSON.stringify({
@@ -281,12 +350,12 @@ export async function DELETE(req) {
       }),
       { status: 200 }
     );
-    } catch (error) {
+  } catch (error) {
     return new Response(
       JSON.stringify({
         message: "An error occurred while deleting the API.",
       }),
       { status: 500 }
     );
-    }
+  }
 }
